Separate part of useAutoConnect to a different hook
Part of the code of useAutoConnect hook was responsible for injecting the wallet
address into url when connecting to a wallet. I separate it to a
useAutoWalletAddressInjectUrl hook, because it wasn't really relate to auto
connect.
