+++ solidity/dashboard/src/App.js
@@ -11,6 +11,8 @@ import * as Icons from "./components/Icons"
 import Footer from "./components/Footer"
 import useSubscribeToConnectorEvents from "./hooks/useSubscribeToConnectorEvents"
 import useAutoConnect from "./hooks/useAutoConnect"
+import useAutoWalletAddressInjectIntoUrl
+  from "./hooks/useAutoWalletAddressInjectIntoUrl";
 
 const App = () => (
   <Provider store={store}>
@@ -28,6 +30,7 @@ const App = () => (
 
 const AppLayout = () => {
   useAutoConnect()
+  useAutoWalletAddressInjectIntoUrl()
   useSubscribeToConnectorEvents()
   return (
     <>
